Refactor usage of WriteBackForTesting
