+++ java/arcs/core/storage/testutil/BUILD
@@ -10,5 +10,8 @@ arcs_kt_jvm_library(
     srcs = glob(["*.kt"]),
     deps = [
         "//java/arcs/core/storage:storage_key",
+        "//java/arcs/core/storage:writeback",
+        "//java/arcs/core/util",
+        "//third_party/kotlin/kotlinx_coroutines:kotlinx_coroutines_test",
     ],
 )
